% Auto-generated: do not edit by hand
\name{dColorDashColor}

\alias{dColorDashColor}

\title{DashColor component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dColorDashColor(id=NULL, Pickerwidth=NULL, disableAlpha=NULL, height=NULL,
presetColors=NULL, value=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{Pickerwidth}{Character. }

\item{disableAlpha}{Logical. }

\item{height}{Character. }

\item{presetColors}{List of characters. }

\item{value}{Character. The value displayed in the input.}

\item{width}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

